+++ constants/index.ts
@@ -2,8 +2,6 @@ export * from './Colors';
 
 export * from './imagesMapWarsh';
 
-export * from './imagesMapMujawad';
-
 export * from './imagesMapHafs';
 
 export * from './slides';