+++ hooks/useImagesArray.ts
@@ -3,7 +3,7 @@ import { useEffect, useRef, useState } from 'react';
 import { Asset } from 'expo-asset';
 import { useAtomValue } from 'jotai/react';
 
-import { imagesMapHafs, imagesMapMujawad, imagesMapWarsh } from '@/constants';
+import { imagesMapHafs, imagesMapWarsh } from '@/constants';
 import useCurrentPage from '@/hooks/useCurrentPage';
 import { mushafRiwaya } from '@/jotai/atoms';
 
@@ -32,9 +32,6 @@ export default function useImagesArray() {
             break;
           case 'warsh':
             imagesMap = imagesMapWarsh;
-            break;
-          case 'mujawad':
-            imagesMap = imagesMapMujawad;
             break;
           default:
             imagesMap = undefined;